+++ src/config.h
@@ -29,6 +29,32 @@
 #define DEFAULT_REFRESH_TIMEOUT 3600
 
 /**
+ * \brief The default HTTP 429 (too many requests) wait time
+ */
+#define DEFAULT_HTTP_WAIT_SEC 5
+
+/**
+ * \brief Data file block size in MB
+ */
+#define DEFAULT_DATA_BLKSZ_MB 8
+
+/**
+ * \brief Data file block size
+ * \details We set it to 1024*1024*8 = 8MiB
+ */
+#define DEFAULT_DATA_BLKSZ (DEFAULT_DATA_BLKSZ_MB * 1024 * 1024)
+
+/**
+ * \brief Maximum segment block count
+ * \details This is set to 128*1024 blocks, which uses 128KB. By default,
+ * this allows the user to store (128*1024)*(8*1024*1024) = 1TB of data
+ */
+#define DEFAULT_MAX_SEGBC (128 * 1024)
+
+#define STR(x) #x
+#define XSTR(x) STR(x)
+
+/**
  * \brief Operation modes
  */
 typedef enum {